ArifurRahmanPython_NBA_Stat_patch
Several patches made for optimization in efficiency regarding graphical representations and increased scalability.  Hopefully on my way to  incorporating statistical measures as far back as the ABA and emergence of the NBA.

Also, working on utilizing the R language, along with Python/HTML/PHP and a bit of JavaScript in order to vividly illustrate the shift of offensive strategy and 3-pt field goal efficiency by analyzing relational statistical differences in attempts and overall percentage increase/decrease.

A fan of this sport has clearly seen a trend which gradually shifted from taking shots at shorter distances with higher probabilities of success to taking long range 3-pt shots with a marginally lower probability of success.  Probability theory and expected outcome computations have proven to favor teams willing to take more 3-pt shots for the opportunity to score.  

A simple explanation can be understood with this example:  Normally a point guard with a fairly